:recycle: Split code of ecell4/python_api/main.cpp
Split main.cpp into core.cpp and main.cpp
