+++ ecell4/python_api/CMakeLists.txt
@@ -1,4 +1,4 @@
-pybind11_add_module(ecell4 main.cpp)
+pybind11_add_module(ecell4 main.cpp core.cpp)
 target_link_libraries(ecell4 PRIVATE
     ecell4-core
     ecell4-bd
