+++ ecell4/python_api/main.cpp
@@ -1,12 +1,7 @@
 #include <pybind11/pybind11.h>
 #include <ecell4/spatiocyte/SpatiocyteWorld.hpp>
-#include "type_caster.hpp"
-#include "core.hpp"
+#include "python_api.hpp"
 #include "world_interface.hpp"
-#include "reaction_rule_descriptor.hpp"
-#include "model.hpp"
-#include "observers.hpp"
-#include "shape.hpp"
 
 namespace py = pybind11;
 
@@ -16,7 +11,7 @@ namespace {
 
     void setup_spatiocyte_module(py::module& m)
     {
-        py::class_<SpatiocyteWorld, WorldInterface, PyWorldImpl<SpatiocyteWorld>>(m, "SpatiocyteWorld")
+        py::class_<SpatiocyteWorld, ecell4::WorldInterface, PyWorldImpl<SpatiocyteWorld>>(m, "SpatiocyteWorld")
             .def(py::init<>());
     }
 
@@ -31,10 +26,5 @@ PYBIND11_MODULE(ecell4, m) {
     py::module m_spatiocyte = m.def_submodule("spatiocyte", "A submodule of ecell4");
 
     setup_module(m);
-    define_world_interface(m);
-    define_reaction_rule_descriptor(m);
-    define_model(m);
-    define_observers(m);
-    define_shape(m);
     setup_spatiocyte_module(m_spatiocyte);
 }
