+++ ecell4/python_api/python_api.hpp
@@ -1,0 +1,19 @@
+#ifndef ECELL4_PYTHON_API_PYTHON_API_HPP
+#define ECELL4_PYTHON_API_PYTHON_API_HPP
+
+#include <pybind11/pybind11.h>
+#include <pybind11/operators.h>
+#include "type_caster.hpp"
+
+namespace ecell4
+{
+
+namespace python_api
+{
+
+void setup_module(pybind11::module& m);
+
+}
+
+}
+#endif /* ECELL4_PYTHON_API_PYTHON_API_HPP */
