+++ ecell4/python_api/shape.hpp
@@ -52,331 +52,34 @@ namespace python_api
         }
     };
 
-    void define_shape(py::module& m)
+    template<class Base>
+    class PyShapeImpl: public PyShape<Base>
     {
-        py::class_<Shape, PyShape<>, boost::shared_ptr<Shape>>(m, "Shape")
-            .def("dimension", [](const Shape& self) { return static_cast<Integer>(self.dimension()); })
-            .def("is_inside", &Shape::is_inside)
-            ;
+    public:
+        using PyShape<Base>::PyShape;
+        using dimension_kind = Shape::dimension_kind;
 
-        py::class_<Surface, Shape, PyShape<Surface>, boost::shared_ptr<Surface>>(m, "Surface")
-            .def("root", &Surface::root)
-            .def(py::pickle(
-                [](const Surface& self)
-                {
-                    return py::make_tuple(self.root());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 1)
-                        throw std::runtime_error("Invalid state");
-                    return Surface(t[0].cast<const boost::shared_ptr<Shape>&>());
-                }
-            ));
+        dimension_kind dimension() const override
+        {
+            PYBIND11_OVERLOAD(dimension_kind, Base, dimension,);
+        }
 
-        py::class_<Union, Shape, PyShape<Union>, boost::shared_ptr<Union>>(m, "Union")
-            .def(py::init<const boost::shared_ptr<Shape>&, const boost::shared_ptr<Shape>&>())
-            .def("surface", &Union::surface)
-            .def("one", &Union::one)
-            .def("another", &Union::another)
-            .def(py::pickle(
-                [](const Union& self)
-                {
-                    return py::make_tuple(self.one(), self.another());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 2)
-                        throw std::runtime_error("Invalid state");
-                    return Union(
-                        t[0].cast<const boost::shared_ptr<Shape>&>(),
-                        t[1].cast<const boost::shared_ptr<Shape>&>()
-                    );
-                }
-            ));
+        Real is_inside(const Real3& coord) const override
+        {
+            PYBIND11_OVERLOAD(Real, Base, is_inside, coord);
+        }
 
-        py::class_<Complement, Shape, PyShape<Complement>, boost::shared_ptr<Complement>>(m, "Complement")
-            .def(py::init<const boost::shared_ptr<Shape>&, const boost::shared_ptr<Shape>&>())
-            .def("surface", &Complement::surface)
-            .def("one", &Complement::one)
-            .def("another", &Complement::another)
-            .def(py::pickle(
-                [](const Complement& self)
-                {
-                    return py::make_tuple(self.one(), self.another());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 2)
-                        throw std::runtime_error("Invalid state");
-                    return Complement(
-                        t[0].cast<const boost::shared_ptr<Shape>&>(),
-                        t[1].cast<const boost::shared_ptr<Shape>&>()
-                    );
-                }
-            ));
+        Real3 draw_position(boost::shared_ptr<RandomNumberGenerator>& rng) const override
+        {
+            PYBIND11_OVERLOAD(Real3, Base, draw_position, rng);
+        }
 
-        py::class_<AffineTransformation, Shape, PyShape<AffineTransformation>,
-            boost::shared_ptr<AffineTransformation>>(m, "AffineTransformation")
-            .def(py::init<>())
-            .def(py::init<const boost::shared_ptr<Shape>&>())
-            .def(py::init<const boost::shared_ptr<Shape>&, const Real3&, const Real3&, const Real3&, const Real3&>())
-            .def("translate", &AffineTransformation::translate)
-            .def("rescale", &AffineTransformation::rescale)
-            .def("xroll", &AffineTransformation::xroll)
-            .def("yroll", &AffineTransformation::yroll)
-            .def("zroll", &AffineTransformation::zroll)
-            .def("surface", &AffineTransformation::surface)
-            .def("root", &AffineTransformation::root)
-            .def("first", &AffineTransformation::first)
-            .def("second", &AffineTransformation::second)
-            .def("third", &AffineTransformation::third)
-            .def("shift", &AffineTransformation::shift)
-            .def(py::pickle(
-                [](const AffineTransformation& self)
-                {
-                    return py::make_tuple(self.root(), self.first(), self.second(), self.third(), self.shift());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 5)
-                        throw std::runtime_error("Invalid state");
-                    return AffineTransformation(
-                        t[0].cast<const boost::shared_ptr<Shape>&>(),
-                        t[1].cast<const Real3&>(),
-                        t[2].cast<const Real3&>(),
-                        t[3].cast<const Real3&>(),
-                        t[4].cast<const Real3&>()
-                    );
-                }
-            ));
+        bool test_AABB(const Real3& l, const Real3& u) const override
+        {
+            PYBIND11_OVERLOAD(bool, Base, test_AABB, l, u);
+        }
+    };
 
-        py::class_<Sphere, Shape, PyShape<Sphere>, boost::shared_ptr<Sphere>>(m, "Sphere")
-            .def(py::init<const Real3&, const Real>())
-            .def("distance", &Sphere::distance)
-            .def("surface", &Sphere::surface)
-            .def("center", &Sphere::center)
-            .def("radius", &Sphere::radius)
-            .def(py::pickle(
-                [](const Sphere& self)
-                {
-                    return py::make_tuple(self.center(), self.radius());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 2)
-                        throw std::runtime_error("Invalid state");
-                    return Sphere(t[0].cast<const Real3&>(), t[1].cast<const Real>());
-                }
-            ));
-
-        py::class_<SphericalSurface, Shape, PyShape<SphericalSurface>,
-            boost::shared_ptr<SphericalSurface>>(m, "SphericalSurface")
-            .def(py::init<const Real3&, const Real>())
-            .def("distance", &SphericalSurface::distance)
-            .def("inside", &SphericalSurface::inside)
-            .def("center", &SphericalSurface::center)
-            .def("radius", &SphericalSurface::radius)
-            .def(py::pickle(
-                [](const SphericalSurface& self)
-                {
-                    return py::make_tuple(self.center(), self.radius());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 2)
-                        throw std::runtime_error("Invalid state");
-                    return SphericalSurface(t[0].cast<const Real3&>(), t[1].cast<const Real>());
-                }
-            ));
-
-        py::class_<Cylinder, Shape, PyShape<Cylinder>, boost::shared_ptr<Cylinder>>(m, "Cylinder")
-            .def(py::init<const Real3&, const Real, const Real3&, const Real>())
-            .def("distance", &Cylinder::distance)
-            .def("surface", &Cylinder::surface)
-            .def("center", &Cylinder::center)
-            .def("axis", &Cylinder::axis)
-            .def("half_height", &Cylinder::half_height)
-            .def(py::pickle(
-                [](const Cylinder& self)
-                {
-                    return py::make_tuple(self.center(), self.radius(), self.axis(), self.half_height());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 4)
-                        throw std::runtime_error("Invalid state");
-                    return Cylinder(
-                        t[0].cast<const Real3&>(),
-                        t[1].cast<const Real>(),
-                        t[2].cast<const Real3&>(),
-                        t[3].cast<const Real>()
-                    );
-                }
-            ));
-
-        py::class_<CylindricalSurface, Shape, PyShape<CylindricalSurface>,
-            boost::shared_ptr<CylindricalSurface>>(m, "CylindricalSurface")
-            .def(py::init<const Real3&, const Real, const Real3&, const Real>())
-            .def("distance", &CylindricalSurface::distance)
-            .def("inside", &CylindricalSurface::inside)
-            .def("center", &CylindricalSurface::center)
-            .def("radius", &CylindricalSurface::radius)
-            .def("axis", &CylindricalSurface::axis)
-            .def("half_height", &CylindricalSurface::half_height)
-            .def(py::pickle(
-                [](const CylindricalSurface& self)
-                {
-                    return py::make_tuple(self.center(), self.radius(), self.axis(), self.half_height());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 4)
-                        throw std::runtime_error("Invalid state");
-                    return CylindricalSurface(
-                        t[0].cast<const Real3&>(),
-                        t[1].cast<const Real>(),
-                        t[2].cast<const Real3&>(),
-                        t[3].cast<const Real>()
-                    );
-                }
-            ));
-
-        py::class_<PlanarSurface, Shape, PyShape<PlanarSurface>,
-            boost::shared_ptr<PlanarSurface>>(m, "PlanarSurface")
-            .def(py::init<const Real3&, const Real3&, const Real3&>())
-            .def("origin", &PlanarSurface::origin)
-            .def("e0", &PlanarSurface::e0)
-            .def("e1", &PlanarSurface::e1)
-            .def("normal", &PlanarSurface::normal)
-            .def(py::pickle(
-                [](const PlanarSurface& self)
-                {
-                    return py::make_tuple(self.origin(), self.e0(), self.e1());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 3)
-                        throw std::runtime_error("Invalid state");
-                    return PlanarSurface(
-                        t[0].cast<const Real3&>(),
-                        t[1].cast<const Real3&>(),
-                        t[2].cast<const Real3&>()
-                    );
-                }
-            ));
-
-        py::class_<Rod, Shape, PyShape<Rod>, boost::shared_ptr<Rod>>(m, "Rod")
-            .def(py::init<const Real&, const Real&>())
-            .def(py::init<const Real&, const Real&, const Real3&>())
-            .def("distance", &Rod::distance)
-            .def("origin", &Rod::origin)
-            .def("length", &Rod::length)
-            .def("radius", &Rod::radius)
-            .def("shift", &Rod::shift)
-            .def("surface", &Rod::surface)
-            .def(py::pickle(
-                [](const Rod& self)
-                {
-                    return py::make_tuple(self.length(), self.radius(), self.origin());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 3)
-                        throw std::runtime_error("Invalid state");
-                    return Rod(
-                        t[0].cast<const Real>(),
-                        t[1].cast<const Real>(),
-                        t[2].cast<const Real3&>()
-                    );
-                }
-            ));
-
-        py::class_<RodSurface, Shape, PyShape<RodSurface>, boost::shared_ptr<RodSurface>>(m, "RodSurface")
-            .def(py::init<const Real&, const Real&>())
-            .def(py::init<const Real&, const Real&, const Real3&>())
-            .def("distance", &RodSurface::distance)
-            .def("origin", &RodSurface::origin)
-            .def("length", &RodSurface::length)
-            .def("radius", &RodSurface::radius)
-            .def("shift", &RodSurface::shift)
-            .def(py::pickle(
-                [](const RodSurface& self)
-                {
-                    return py::make_tuple(self.length(), self.radius(), self.origin());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 3)
-                        throw std::runtime_error("Invalid state");
-                    return RodSurface(
-                        t[0].cast<const Real>(),
-                        t[1].cast<const Real>(),
-                        t[2].cast<const Real3&>()
-                    );
-                }
-            ));
-
-        py::class_<AABB, Shape, PyShape<AABB>, boost::shared_ptr<AABB>>(m, "AABB")
-            .def(py::init<const Real3&, const Real3&>())
-            .def("distance", &AABB::distance)
-            .def("upper", &AABB::upper)
-            .def("lower", &AABB::lower)
-            .def("surface", &AABB::surface)
-            .def(py::pickle(
-                [](const AABB& self)
-                {
-                    return py::make_tuple(self.lower(), self.upper());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 2)
-                        throw std::runtime_error("Invalid state");
-                    return AABB(
-                        t[0].cast<const Real3&>(),
-                        t[1].cast<const Real3&>()
-                    );
-                }
-            ));
-
-        py::class_<MeshSurface, Shape, PyShape<MeshSurface>, boost::shared_ptr<MeshSurface>>(m, "MeshSurface")
-            .def(py::init<const std::string, const Real3&>())
-            .def("filename", &MeshSurface::filename)
-            .def("edge_lengths", &MeshSurface::edge_lengths)
-            .def(py::pickle(
-                [](const MeshSurface& self)
-                {
-                    return py::make_tuple(self.filename(), self.edge_lengths());
-                },
-                [](py::tuple t)
-                {
-                    if (t.size() != 2)
-                        throw std::runtime_error("Invalid state");
-                    return MeshSurface(
-                        t[0].cast<const std::string>(),
-                        t[1].cast<const Real3&>()
-                    );
-                }
-            ));
-
-        m.def("create_x_plane",
-            [](Real x)
-            {
-                return PlanarSurface(Real3(x, 0, 0), Real3(0, 1, 0), Real3(0, 0, 1));
-            });
-
-        m.def("create_y_plane",
-            [](Real y)
-            {
-                return PlanarSurface(Real3(0, y, 0), Real3(1, 0, 0), Real3(0, 0, 1));
-            });
-
-        m.def("create_z_plane",
-            [](Real z)
-            {
-                return PlanarSurface(Real3(0, 0, z), Real3(1, 0, 0), Real3(0, 1, 0));
-            });
-    }
 }
 
 }
